% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_tools.R
\name{alignment_data_to_data_frame}
\alias{alignment_data_to_data_frame}
\title{transform alignment_data list into data.frame}
\usage{
alignment_data_to_data_frame(x)
}
\arguments{
\item{x}{alignment_data list}
}
\description{
transform alignment_data list into data.frame
}
\keyword{internal}

