% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{choose_options}
\alias{choose_options}
\title{(choose from a number of pre-sorted options)
takes a vector pair of toki1 / toki2 and a vector pair of res_token_i_1 /
res_token_i_2 and chooses so that each 1st and exh 2nd value only is used
where res_token_i_x identiefies already used items.}
\usage{
choose_options(toki1, toki2, res_token_i_1, res_token_i_2)
}
\arguments{
\item{toki1}{first number of number pair to choose from}

\item{toki2}{second number of number pair to choose from}

\item{res_token_i_1}{already used first numbers}

\item{res_token_i_2}{already used second numbers
// @keywords internal}
}
\description{
(choose from a number of pre-sorted options)
takes a vector pair of toki1 / toki2 and a vector pair of res_token_i_1 /
res_token_i_2 and chooses so that each 1st and exh 2nd value only is used
where res_token_i_x identiefies already used items.
}

