% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffr.R
\name{diffr}
\alias{diffr}
\title{FUNCTION_TITLE}
\usage{
diffr(text1 = NULL, text2 = NULL, tokenizer = NULL, ignore = NULL,
  clean = NULL, distance = c("lv", "osa", "dl", "hamming", "lcs", "qgram",
  "cosine", "jaccard", "jw", "soundex"), ...)
}
\arguments{
\item{text1}{first text}

\item{text2}{second text}

\item{tokenizer}{defaults to NULL which will trigger linewise tokenization;
accepts a function that turns a text into a token data frame;
a token data frame has at least three columns:
from (first character of token),
to (last character of token)
token (the token)}

\item{ignore}{defaults to NULL which means that nothing is ignored;
function that accepts a token data frame (see above) and returns a
possibly subseted data frame of hte same form}

\item{clean}{defaults to NULL which means that nothing cleaned; accepts a
function that takes a vector of tokens and returns a vector of same
length - potentially clean up}

\item{distance}{defaults to Levenshtein ("lv"); see \link[stringdist]{amatch},
\link[stringdist]{stringdist-metrics}, \link[stringdist]{stringdist}}

\item{...}{further arguments passed through to distance function}
}
\value{
dataframe with tokens aligned according to distance
}
\description{
FUNCTION_DESCRIPTION
}

