% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_mat_absolute}
\alias{dist_mat_absolute}
\title{(function to calculate distance matrix of integers)
takes vector of size n and vector of size m and gives back matrix of n rows and m columns}
\usage{
dist_mat_absolute(x, y)
}
\arguments{
\item{x}{a vector of type numeric}

\item{y}{a vector of type numeric}
}
\description{
(function to calculate distance matrix of integers)
takes vector of size n and vector of size m and gives back matrix of n rows and m columns
}
\keyword{internal}

