% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_align.R
\docType{class}
\name{dp_align}
\alias{dp_align}
\title{class for dp_align}
\format{\code{\link{R6Class}} object.}
\usage{
dp_align
}
\value{
Object of \code{\link{dp_align}}
}
\description{
class for dp_align
}
\seealso{
\code{\link{diffrproject}}
}
\keyword{data}

