% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_export.R
\docType{class}
\name{dp_export}
\alias{dp_export}
\title{R6 class - linking text and data}
\format{\code{\link{R6Class}} object.}
\usage{
dp_export
}
\value{
Object of \code{\link{R6Class}}
}
\description{
R6 class - linking text and data
}
\seealso{
\code{\link{diffrproject}}
}
\keyword{data}

