% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_inherit.R
\docType{class}
\name{dp_inherit}
\alias{dp_inherit}
\title{class for dp_inherit}
\format{\code{\link{R6Class}} object.}
\usage{
dp_inherit
}
\value{
Object of \code{\link{dp_align}}
}
\description{
class for dp_inherit
}
\seealso{
\code{\link{diffrproject}}
}
\keyword{data}

