% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_loadsave.R
\docType{class}
\name{dp_loadsave}
\alias{dp_loadsave}
\title{class for dp_base}
\format{\code{\link{R6Class}} object.}
\usage{
dp_loadsave
}
\value{
Object of \code{\link{dp_loadsave}}
}
\description{
class for dp_base
}
\seealso{
\code{\link{diffrproject}}
}
\keyword{data}

