% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_list_item}
\alias{get_list_item}
\title{extract specific item from each list element}
\usage{
get_list_item(l, item, unlist = TRUE)
}
\arguments{
\item{l}{list}

\item{item}{name or index of item to extract}
}
\description{
extract specific item from each list element
}

