% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_private}
\alias{get_private}
\title{accessing private from R6 object}
\source{
http://stackoverflow.com/a/38578080/1144966
}
\usage{
get_private(x)
}
\arguments{
\item{x}{R6 object to access private from}
}
\description{
accessing private from R6 object
}

