% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is_duplicate}
\alias{is_duplicate}
\title{checking if value is duplicated in set}
\usage{
is_duplicate(x)
}
\arguments{
\item{x}{vector to check}
}
\description{
checking if value is duplicated in set
}
\keyword{internal}

