% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is_minimum}
\alias{is_minimum}
\title{which are minima in vector}
\usage{
is_minimum(x, unique = FALSE)
}
\arguments{
\item{x}{vector to check}

\item{unique}{defaults to false}
}
\description{
which are minima in vector
}
\keyword{internal}

