% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moc.R
\name{moc}
\alias{moc}
\title{stub}
\usage{
moc(text1 = NULL, text2 = NULL, tokenizer = function(text) {    
  text_tokenize_lines(text) }, ignore = function(...) {     FALSE },
  clean = function(token) {     token }, distance = function(token1, token2)
  {     matrix(0, nrow = length(token1), ncol = length(token2)) },
  alignment = function(m) { })
}
\description{
stub
}
\keyword{internal}

