% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moc_helper.R
\name{moc_helper_easy_matches}
\alias{moc_helper_easy_matches}
\title{easy matches 1}
\usage{
moc_helper_easy_matches(tt1, tt2, res, type = c(1, 2), fullreturn = TRUE)
}
\arguments{
\item{tt1}{tokenized text number 1}

\item{tt2}{tokenized text number 2}
}
\description{
method of comparison helper function
}
\keyword{internal}

