% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moc_helper.R
\name{moc_helper_get_options_ordered_by_dist}
\alias{moc_helper_get_options_ordered_by_dist}
\title{get options for machtches}
\usage{
moc_helper_get_options_ordered_by_dist(tt1, tt2, res)
}
\arguments{
\item{tt1}{tokenized text number 1}

\item{tt2}{tokenized text number 2}

\item{res}{data.frame of already matched}
}
\description{
method of comparison helper function
}

