% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moc_helper.R
\name{moc_helper_trivial_matches}
\alias{moc_helper_trivial_matches}
\title{trivial matches}
\usage{
moc_helper_trivial_matches(tt1, tt2)
}
\arguments{
\item{tt1}{tokenized text number 1}

\item{tt2}{tokenized text number 2}
}
\description{
method of comparison helper function
}
\keyword{internal}

