% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_tools.R
\name{rtext_char_data_to_data_frame}
\alias{rtext_char_data_to_data_frame}
\title{transform rtext text data into a data.frame}
\usage{
rtext_char_data_to_data_frame(x)
}
\arguments{
\item{x}{rtext object}
}
\description{
transform rtext text data into a data.frame
}
\keyword{internal}

