% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_tools.R
\name{sort_alignment}
\alias{sort_alignment}
\title{function sorting alignment data according to token index}
\usage{
sort_alignment(x, ti1 = NULL, ti2 = NULL, first = TRUE)
}
\arguments{
\item{x}{data.frame to be sorted}

\item{ti1}{either NULL (default): first column of x is used as first token
index for sorting; a character vector specifying the column to be used
as first token index; or a numeric vector of length nrow(x) to be use
as first token index}

\item{ti2}{either NULL (default): second column of x is used as second token
index for sorting; a character vector specifying the column to be used
as second token index; or a numeric vector of length nrow(x) to be use
as second token index}

\item{first}{should first text or second text be given priority}
}
\description{
function sorting alignment data according to token index
}

