% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moc_helper.R
\name{split_tt_by_length}
\alias{split_tt_by_length}
\title{splitting a tokenized text}
\usage{
split_tt_by_length(tt)
}
\arguments{
\item{tt}{tokenized text}
}
\description{
splitting a tokenized text
}
\keyword{internal}

