% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_tools.R
\name{text_add_worker}
\alias{text_add_worker}
\title{function adding rtext objects to diffrprojects}
\usage{
text_add_worker(self, rtext = NULL, name = NULL)
}
\arguments{
\item{self}{an object of class dp}

\item{rtext}{an object of class rtext}

\item{name}{an optional name for the text to stick to the text within the
diffrproject corpus - if none is supplied the function will try to
infere a reasonable name from the rtext$text_file field, if that is
not given it will get the name noname_x where x is a running integer}
}
\description{
function adding rtext objects to diffrprojects
}
\keyword{internal}

