% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texts.R
\docType{data}
\name{text_version_1}
\alias{text_version_1}
\title{text_version_1 a first version of a text}
\format{An object of class \code{character} of length 1.}
\source{
Source of Text: Diff. (2014, August 26). In Wikipedia, The Free Encyclopedia. Retrieved 10:14, September 24, 2014, from http://en.wikipedia.org/w/index.php?title=Diff&oldid=622929855
}
\usage{
text_version_1
}
\description{
text_version_1 a first version of a text
}
\keyword{datasets}

