% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{which_dist_min_absolute}
\alias{which_dist_min_absolute}
\title{(function to calculate minimum and position of minimum)
takes vector of size n and vector of size m and gives back list with
vectors of size n (minimum distance and location of minimum in y)}
\usage{
which_dist_min_absolute(x, y)
}
\arguments{
\item{x}{a vector of type integer}

\item{y}{a vector of type integer}
}
\description{
(function to calculate minimum and position of minimum)
takes vector of size n and vector of size m and gives back list with
vectors of size n (minimum distance and location of minimum in y)
}
\keyword{internal}

