% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R, R/dp_align.R, R/dp_base.R
\docType{class}
\name{diffrproject}
\alias{diffrproject}
\alias{dp_align}
\alias{dp_base}
\title{class for diffrproject}
\format{\code{\link{R6Class}} object.}
\usage{
diffrproject

dp_align

dp_base
}
\value{
Object of \code{\link{diffrproject}}

Object of \code{\link{dp_align}}

Object of \code{\link{dp_base}}
}
\description{
class for diffrproject

class for dp_align

class for dp_base
}
\keyword{data}

