% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R, R/dp_align.R, R/dp_base.R,
%   R/dp_inherit.R, R/dp_loadsave.R
\docType{class}
\name{diffrproject}
\alias{diffrproject}
\alias{dp_align}
\alias{dp_base}
\alias{dp_inherit}
\alias{dp_loadsave}
\title{class for diffrproject}
\format{\code{\link{R6Class}} object.}
\usage{
diffrproject

dp_align

dp_base

dp_inherit

dp_loadsave
}
\value{
Object of \code{\link{diffrproject}}

Object of \code{\link{dp_align}}

Object of \code{\link{dp_base}}

Object of \code{\link{dp_align}}

Object of \code{\link{dp_loadsave}}
}
\description{
class for diffrproject

class for dp_align

class for dp_base

class for dp_inherit

class for dp_base
}
\keyword{data}

