% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_tools.R
\name{write_numerous_parts_to_table}
\alias{write_numerous_parts_to_table}
\title{function writing numerous parts of table to database}
\usage{
write_numerous_parts_to_table(x, con, table_name, meta = data.frame())
}
\arguments{
\item{x}{parts to be written}

\item{con}{connection to database}

\item{table_name}{of the table}

\item{meta}{additional information to be attachesd to table parts}
}
\description{
function writing numerous parts of table to database
}

